Abstract player functionality into one folder
