+++ client/schedule/index.js
@@ -1,0 +1,5 @@
+import React, { Component } from 'react';
+
+export const Schedule = () => (
+	<h2> Game Schedule </h2>
+)
