+++ client/src/actions/index.js
@@ -12,24 +12,6 @@ export const FILTER_PLAYERS = 'FILTER_PLAYERS';
 export const GET_ONE_PLAYER_MODAL = 'GET_ONE_PLAYER_MODAL';
 export const CHANGE_SEARCH_TEXT = 'CHANGE_SEARCH_TEXT';
 
-export function fetchRoster(league_key) {
-  const request = axios.get(`/roster/${league_key}`).then((team) => {
-    const playerId = [];
-    for (let i = 0; i < team.data.length; i++) {
-      playerId.push(Number(team.data[i].player_id));
-    }
-    const statsRequest = axios.post('/api/getPlayersByIds', { playerId })
-      .then((stats) => {
-        return { players: team.data, stats: stats.data[0] };
-      });
-
-    return statsRequest;
-  });
-  return {
-    type: FETCH_ROSTER,
-    payload: request,
-  };
-}
 
 // Toggles shadow navigation for off-canvas navigation bar
 export function toggleNavigation() {
