+++ client/src/player/ClickableHeading.jsx
